encoder: allow re-fetching containers nested under a keyed container
This allows types declaring their own Encodable conformance to call
nestedUnkeyedContainer(forKey:) or nestedContainer(keyedBy:forKey:)
without having to keep track of whether it has been called before for
the exact CodingPath, because it will return the previously made
container if one exists.

The old behavior is that it would overwrite the existing container at
that key.